requirements.txt
chainlit
chromadb
langchain
langchain-community
langchain-text-splitters
langchain-ollama
pypdf
python-dotenv
pytesseract
pillow
